handle non-damage flushes
When weston internally damages a surface, we don't get a damage-commit
event in timeline. Cursors hit this somehow.

Do not ignore non-damage flushes, but create an update item for them
with unset damage timestamp.

Modify update_to_svg() to deal with all combinations of damage, flush,
or vblank timestamps missing.